pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-signals"
version = "0.1.0"
description = "Institutional-grade AI trading signal generator"
requires-python = ">=3.10"
authors = [
    {name = "Nirav Polara", email = "nirav@example.com"}
]

dependencies = [
    # Data & Time Series
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "TA-Lib>=0.4.28",
    
    # Database
    "SQLAlchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.12.0",
    
    # API & Async
    "requests>=2.31.0",
    "httpx>=0.24.0",
    "aiohttp>=3.9.0",
    "yfinance>=0.2.32",
    
    # Task Queue
    "celery>=5.3.0",
    "redis>=5.0.0",
    
    # NLP & LLM
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "openai>=1.3.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "pinecone-client>=2.2.0",
    
    # Frontend
    "dash>=2.14.0",
    "dash-bootstrap-components>=1.4.0",
    "plotly>=5.17.0",
    
    # Backtesting
    "vectorbt>=0.25.0",
    
    # Configuration & Logging
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=23.1.0",
    
    # Authentication
    "bcrypt>=4.0.0",
    "flask-login>=0.6.2",
    "flask-cors>=4.0.0",
    
    # Payments (Phase 1.2)
    "stripe>=7.0.0",
    
    # Email (Phase 2.2)
    "sendgrid>=6.10.0",
    
    # Monitoring
    "sentry-sdk>=1.38.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "jupyter>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
